+++ src/Context/index.js
@@ -1,0 +1,2 @@
+export { useReducerContext, ReducerProvider } from "./Reducer.context";
+export { useThemeContext, ThemeProvider } from "./Theme.context";
