+++ src/Pages/ListingPage/ListingPageComponents/index.js
@@ -1,0 +1,2 @@
+export { CategoryBar } from "./CategoryBar/CategoryBar";
+export { VideoListing } from "./VideoListing/VideoListing";
