+++ src/Pages/index.js
@@ -1,1 +1,2 @@
 export { LandingPage } from "./LandingPage/LandingPage";
+export { ListingPage } from "./ListingPage/ListingPage";
